requirements.txt
accelerate==1.1.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
beartype==0.14.1
beautifulsoup4==4.12.3
better-abc==0.0.3
bleach==6.2.0
blis==0.7.7
Bottleneck==1.3.4
brotlipy==0.7.0
catalogue==2.0.7
certifi==2022.12.7
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
cmake==3.31.1
comm==0.2.2
conda-pack==0.8.1
contourpy==1.3.1
cryptography==37.0.1
cycler==0.12.1
cymem==2.0.6
datasets==3.1.0
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.8.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fairscale==0.4.13
fancy-einsum==0.0.3
fastjsonschema==2.21.1
filelock==3.9.0
fonttools==4.55.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2023.5.0
gitdb==4.0.9
GitPython==3.1.27
gmpy2==2.1.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
hydra-core==1.2.0
idna==3.3
ipykernel==6.29.5
ipython==8.30.0
ipywidgets==8.1.5
isoduration==20.11.0
jaxtyping==0.2.36
jedi==0.19.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
langcodes==3.3.0
lit==18.1.8
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mpmath==1.2.1
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.11
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.4
notebook==7.2.2
notebook_shim==0.2.4
numexpr==2.8.1
numpy==2.1.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.2.3
openai==0.20.0
openpyxl==3.0.10
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-stubs==1.2.0.61
pandocfilters==1.5.1
parso==0.8.4
pathtools==0.1.2
pathy==0.6.1
peft==0.3.0
pexpect==4.9.0
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==4.3.6
preshed==3.0.9
prometheus_client==0.21.1
promise==2.3
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.18.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.1
PyYAML==6.0
pyzmq==26.2.0
referencing==0.35.1
regex==2022.6.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.0
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.1.97
sentry-sdk==2.19.0
setproctitle==1.2.3
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.8
stack-data==0.6.3
sympy==1.11.1
terminado==0.18.1
thinc==8.0.17
threadpoolctl==3.1.0
tinycss2==1.4.0
tokenizers==0.20.3
tomli==2.2.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformer-lens==2.9.1
transformers==4.46.3
triton==2.0.0
typeguard==4.4.1
typer==0.4.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.2.3
wandb==0.18.7
wasabi==0.9.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.13
xxhash==3.5.0
yarg==0.1.9
yarl==1.18.3
